Significant code refactoring.  Details follow.
The API for all of the functions in level.c and build.c has changed---for the
better, I'd say.  This was prompted by a desire to start implementing support
for '!' patterns in level files.  There are only two of them right now, and I
haven't actually even tested them, BUT, it's now fairly easy to add them.
Even more importantly, from the new position, it's easier to make a flexible
system for changing what patterns we support, using function pointers or
something.

The one part of ! patterns that HAS been tested is that, if there is an
optional pattern that isn't supported, we succesfully ignore it.

Building a graph is now easier: call level_parse to get NULL-terminated lists
of SavedTiles and ConfigOptions, then call board_assemble_full to get a Board.
I might make a wrapper that does those both at once: it seems handy.

No longer represent the zero-step path as a NULL pointer.  This wasn't a
problem before, but now a NULL pointer is used to figure out if we're in the
NULL termination of the list---before, we passed around the list's length.
Instead, use a pointer to a NULL byte, which is to say, a pointer to the empty
string.  (Which is more intuitive anyways.)

Audits to either the initial parsed information or to the final board state
can be added in audit.c.  Both audits currently return 1 (success) right now,
without doing anything.